+++ gcp-test/tests/gcp_compute_target_ssl_proxy/test-get-query.sql
@@ -1,0 +1,3 @@
+select name, description, kind, proxy_header, service, self_link, ssl_certificates, project, location
+from gcp.gcp_compute_target_ssl_proxy
+where name = '{{ resourceName }}';