+++ gcp-test/tests/gcp_dns_policy/test-get-query.sql
@@ -1,0 +1,11 @@
+select
+  name,
+  description,
+  kind,
+  enable_inbound_forwarding,
+  enable_logging,
+  project,
+  networks,
+  target_name_servers
+from gcp.gcp_dns_policy
+where name = '{{ resourceName }}';