+++ gcp-test/tests/gcp_kms_key/test-list-query.sql
@@ -1,0 +1,3 @@
+select name, rotation_period
+from gcp.gcp_kms_key
+where akas::text = '["{{ output.resource_aka.value }}"]';