+++ gcp-test/tests/gcp_kms_key_ring/test-list-query.sql
@@ -1,0 +1,3 @@
+select name
+from gcp.gcp_kms_key_ring
+where akas::text = '["{{ output.resource_aka.value }}"]'