+++ gcp-test/tests/gcp_logging_metric/test-hydrate-query.sql
@@ -1,3 +1,3 @@
-select name, metric_descriptor_display_name, metric_descriptor_metric_kind, metric_descriptor_unit, metric_descriptor_value_type, metric_descriptor_labels, filter, description, value_extractor, label_extractors
+select name, metric_descriptor_display_name, metric_descriptor_metric_kind, metric_descriptor_type, metric_descriptor_unit, metric_descriptor_value_type, metric_descriptor_labels, filter, description, value_extractor, label_extractors
 from gcp.gcp_logging_metric
 where name = '{{resourceName}}'