+++ gcp-test/tests/gcp_logging_sink/test-get-query.sql
@@ -1,3 +1,3 @@
-select name, destination, disabled, filter, description, include_children, project, location
+select name, destination, disabled, filter, description, include_children, self_link, project, location
 from gcp.gcp_logging_sink
 where name = '{{ resourceName }}';