+++ gcp-test/tests/gcp_logging_sink/test-hydrate-query.sql
@@ -1,3 +1,3 @@
 select name, destination, disabled, filter, description, include_children
 from gcp.gcp_logging_sink
-where name = '{{resourceName}}'
+where name = '{{resourceName}}';