+++ gcp-test/tests/gcp_project/test-notfound-query.sql
@@ -1,0 +1,7 @@
+select
+  name,
+  title
+from
+  gcp_project
+where
+  name = 'dummy-{{ resourceName }}';